+++ src/vector2.c
@@ -1,6 +1,6 @@
 #include "include/vector2.h"
 
-void vector2_add(struct vector2 *vector, struct vector2 v1, struct vector2 v2, float k)
+void vector2_add(struct vector2 *vector, struct vector2 v1, struct vector2 v2, double k)
 {
     vector->x = v1.x + v2.x * k;
     vector->y = v1.y + v2.y * k;
@@ -12,7 +12,7 @@ void vector2_multiply(struct vector2 *vector, struct vector2 v1, struct vector2 
     vector->y = v1.y * v2.y;
 }
 
-void vector2_scale(struct vector2 *vector, float k)
+void vector2_scale(struct vector2 *vector, double k)
 {
     vector->x *= k;
     vector->y *= k;
@@ -25,7 +25,7 @@ void vector2_cpy(struct vector2 *v1, struct vector2 v2)
 }
 
 /*
-float vector2_len(struct vector2 v)
+double vector2_len(struct vector2 v)
 {
     return sqrt(v.x * v.x + v.y * v.y);
 }
@@ -34,7 +34,7 @@ float vector2_len(struct vector2 v)
 /*
 void vector2_normalize(struct vector2 *v)
 {
-    float len = vector2_len(*v);
+    double len = vector2_len(*v);
     vector2_scale(v, 1/len);
 }
 */
